requirements.txt
paramiko~=2.11.0
Click~=8.1.3
pyyaml~=6.0
setuptools~=65.5.0
wheel